parity: tighter SSM recurrent / gated-rmsnorm path against llama.cpp
Brings layer 0 to 11/16 bit-identical checkpoints (was 11/16 before
this commit too, but the residual drift at attn_output / mix / l_out
is now 3-9 head ULPs instead of 4-20). Self-test still passes.

- llm.c llm_forward_ssm step 9 (recurrent state update):
  - kv_mem and out reductions now accumulate in fp64. ggml's
    sum_rows uses ggml_float (= double) when Accelerate is not
    linked, and vDSP_sve when it is - both fp64-equivalent. Scalar
    fp32 was 1-6 ULPs off per element. With fp64 some elements get
    closer, some stay; the residual is genuinely ggml-internal
    reduction order (vDSP_sve is Apple's scalar API surface, not
    reproducible without linking Accelerate).

- llm.c llm_forward_ssm step 10 (gated RMSNorm + z gate):
  - SSQ accumulator: (double)(yh[v] * yh[v]) instead of
    (double)yh[v] * (double)yh[v]. Matches ggml's RMS_NORM kernel.
  - z is now SiLU'd via neon_silu_vec_f32 (NEON polynomial expf)
    rather than scalar libm `x / (1 + expf(-x))`. Matches ggml's
    GGML_OP_SILU bit-exactly.

- tensor.c tensor_rms_norm:
  - Same (double)(x*x) accumulator pattern as ggml. Self-test still
    produces identical argmax (logit values are bit-equal too).

Diagnostic table for the "Hello" greedy parity:

  ✓ inp_embd            ✓ attn_norm-0       ✓ attn_qkv-0
  ✓ conv_silu-0         ✓ Q_l2norm-0        ✓ K_l2norm-0
  ✓ alpha-0             ✓ beta-0            ✓ a_softplus-0
  ✓ beta_in-0           ✓ g_in-0
  ✗ attn_output-0       (3 head ULP drift, recurrent state sum)
  ✗ y_norm-0            (2 head ULP drift)
  ✗ mix-0               (9 head ULP drift, compounded)
  ✗ ffn_norm-0          (16 head ULP)
  ✗ l_out-0             (152 head ULP, layer-0 final residual)

+++ llm/tensor.c
@@ -326,7 +326,7 @@ struct tensor * tensor_rms_norm(struct tensor * x,
         float * or_ = out->data + t * dim;
         double ssq = 0.0;
         for (int64_t i = 0; i < dim; i++) {
-            ssq += (double)xr[i] * (double)xr[i];
+            ssq += (double)(xr[i] * xr[i]);
         }
         float rscale = 1.0f / sqrtf((float)(ssq / (double)dim) + eps);
         for (int64_t i = 0; i < dim; i++) {
